회원가입시 Default 알람 타임 설정 특정 약물 조회 api 추가 특정 질병 조회 api 추가 회원가입 시 user_medication 테이블 <-> user_prescription 관계 설정으로 인한 회원 탈퇴 오류 수정 api 호출시 Token == 엔드포인트username 일치 확인(액세스 토큰 사용자 api 전체 테스트 후 정상 작동 확인

+++ Altari/src/main/java/org/zerock/Altari/controller/RegisterController.java
@@ -13,9 +13,12 @@ import org.zerock.Altari.dto.RegisterDTO;
 import org.zerock.Altari.dto.UserProfileDTO;
 import org.zerock.Altari.entity.UserEntity;
 import org.zerock.Altari.entity.UserProfileEntity;
+import org.zerock.Altari.exception.EntityNotMatchedException;
+import org.zerock.Altari.security.util.JWTUtil;
 import org.zerock.Altari.service.RegisterService;
 import org.zerock.Altari.service.UserService;
 
+import java.io.UnsupportedEncodingException;
 import java.util.Map;
 import java.util.NoSuchElementException;
 
@@ -26,6 +29,7 @@ import java.util.NoSuchElementException;
 public class RegisterController {
 
     private final RegisterService registerService;
+    private final JWTUtil jwtUtil;
 
 
     @PostMapping("/register")
@@ -41,7 +45,18 @@ public class RegisterController {
     }
 
     @DeleteMapping("/delete/{username}")
-    public ResponseEntity<String> deleteUser(@PathVariable String username) {
+    public ResponseEntity<String> deleteUser(@PathVariable String username,
+                                             @RequestHeader("Authorization") String accessToken) throws UnsupportedEncodingException {
+
+        UserEntity userToken = jwtUtil.getUsernameFromToken(accessToken);
+        UserEntity user = new UserEntity(username);
+        String tokenUsername = userToken.getUsername();
+        String entityUsername = user.getUsername();
+
+        // 3. userToken과 user가 다르면 예외 처리
+        if (!tokenUsername.equals(entityUsername)) {
+            throw new EntityNotMatchedException("권한이 없습니다.");
+        }
         try {
             registerService.deleteUser(username);
             return ResponseEntity.ok("회원 탈퇴가 완료되었습니다.");
